cmd/ld: fix some 64-bit issues A few places in the linker pushed 64-bit values through 32-bit holes, including in relocation. Clean them up, and check for a few other overflows as well. Tests to follow.
R=dsymonds
CC=gobot, golang-dev
https://golang.org/cl/9032043

+++ src/cmd/ld/ldmacho.c
@@ -804,9 +804,9 @@ ldmacho(Biobuf *f, char *pkg, int64 len, char *pn)
 				//
 				// [For future reference, see Darwin's /usr/include/mach-o/x86_64/reloc.h]
 				secaddr = c->seg.sect[rel->symnum-1].addr;
-				rp->add = e->e32(s->p+rp->off) + rp->off + 4 - secaddr;
+				rp->add = (int32)e->e32(s->p+rp->off) + rp->off + 4 - secaddr;
 			} else
-				rp->add = e->e32(s->p+rp->off);
+				rp->add = (int32)e->e32(s->p+rp->off);
 
 			// For i386 Mach-O PC-relative, the addend is written such that
 			// it *is* the PC being subtracted.  Use that to make
